% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sf.R
\name{add_area}
\alias{add_area}
\title{Add area}
\usage{
add_area(x, col_name = ".area", drop_units = FALSE)
}
\arguments{
\item{x}{an sf object.}

\item{col_name}{a character (length 1). A name for the new column.}

\item{drop_units}{a logical. Should the area units be droppped?}
}
\value{
the given sf object with an additional column.
}
\description{
Add a column with the area to the given object.
}
