% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sf.R
\name{add_centroids_coords}
\alias{add_centroids_coords}
\title{Add centroid coordinates}
\usage{
add_centroids_coords(x, col_names = c("X", "Y"))
}
\arguments{
\item{x}{an sf object.}

\item{col_names}{a character (length 2). Names for the new XY columns.}
}
\value{
the given sf object with additional columns.
}
\description{
Add columns with the coordinates of its centroids to the given object.
}
