% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sf.R
\name{add_length}
\alias{add_length}
\title{Add length}
\usage{
add_length(x, col_name = ".length", drop_units = FALSE)
}
\arguments{
\item{x}{an sf object.}

\item{col_name}{a character (length 1). A name for the new column.}

\item{drop_units}{a logical. Should the area units be droppped?}
}
\value{
the given sf object with an additional column.
}
\description{
Add a column with lengths to the given object.
}
