% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregate_geom}
\alias{aggregate_geom}
\title{Aggregate vector data using a grid}
\usage{
aggregate_geom(x, by, grid, grid_id, funs, ...)
}
\arguments{
\item{x}{an sf object.}

\item{by}{a character(1). A column name in \code{x}.}

\item{grid}{an sf object (polygon). A grid used for aggregating \code{x}.}

\item{grid_id}{a character(1). A column name in \code{grid} with identifiers for
each cell in \code{grid}.}

\item{funs}{a character. Name of funcions for aggregating the data in \code{x}
that intersects with each cell in \code{grid}.}

\item{...}{additional parameters of \code{funs}.}
}
\value{
\code{grid} with additional attributes.
}
\description{
This function estimates the geometric properties (number, length, and area)
resulting from the intersection of the given vector data and a grid.
}
