% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregate_geom_categ}
\alias{aggregate_geom_categ}
\title{Aggregate geometries by category}
\usage{
aggregate_geom_categ(categ, data_sf_ls, grid, grid_id, funs, ...)
}
\arguments{
\item{categ}{a character(1). Name of a category in the elements of \code{data_sf_ls}.}

\item{data_sf_ls}{a list of sf objects. The data in this list is meant to be
aggregated using the given \code{grid}.}

\item{grid}{an sf object (polygons). Grid used to aggregate data.}

\item{grid_id}{a character(1). Name of a column in \code{grid} that identifies
each cell in \code{grid}.}

\item{funs}{a character. Name of aggregation functions.}

\item{...}{parameters passed to funs.}
}
\value{
a list of data frames with the results of the aggrgation of the
spatial attributes of the data in \code{data_sf_ls} using \code{grid}.
}
\description{
Helper function called from \code{aggregate_geom}.
}
