% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_raster.R
\name{aggregate_raster}
\alias{aggregate_raster}
\title{Aggregate raster data}
\usage{
aggregate_raster(x, by, grid, grid_id, funs, ...)
}
\arguments{
\item{x}{a terra object.}

\item{by}{a character(1). A column name in \code{x}.}

\item{grid}{A grid (polygons) used to aggregate \code{x}.}

\item{grid_id}{a character(1). A column name in \code{grid} with an unique
identifier for each row.}

\item{funs}{a character. Functions names used to aggregate the data in \code{x}
corresponding to each cell in \code{grid}.}

\item{...}{additional parameters passed to funs.}
}
\value{
\code{grid} (sf) with additional columns.
}
\description{
Aggregate raster data using simple statistics and a grid (vector).
}
