% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregate_vector}
\alias{aggregate_vector}
\title{Aggregate vector data}
\usage{
aggregate_vector(x, by, grid, grid_id, funs, ...)
}
\arguments{
\item{x}{an sf object.}

\item{by}{a character(1). A column name in \code{x} with a category for each row.}

\item{grid}{and sf object (polygon). A grid used to aggregate \code{x}.}

\item{grid_id}{a character(1). A column name in \code{grid} with an unique
identifier for each row.}

\item{funs}{a character. Functions names used to aggregate the data in \code{x}
corresponding to each cell in \code{grid}.}

\item{...}{additional parameters passed to funs.}
}
\description{
Aggregate vector data (points, lines, polygons) and their attributes using
simple statistics, a grid, and the data categories.
}
