% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{clean_names}
\alias{clean_names}
\title{Clean the names in a data.frame}
\usage{
clean_names(.data, unique = FALSE, to_lower = FALSE, to_upper = FALSE)
}
\arguments{
\item{.data}{a data.frame or a character.}

\item{unique}{a logical indicating is the produced names should be unique.}

\item{to_lower, to_upper}{Should the results be converted to a single case?}
}
\value{
a data.frame or a character.
}
\description{
Parse a character or a data.frame's names
}
\seealso{
This function was adapted from the blog post called "Clean,
Consistent Column Names" by William Doane publishe in R-bloggers July 7,
2019.
}
