% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sf.R
\name{get_center}
\alias{get_center}
\title{Get center coordinates}
\usage{
get_center(x)
}
\arguments{
\item{x}{an sf or terra object.}
}
\value{
a named vector.
}
\description{
Get the coordinates of the center of the extent of the given  object (sf or
terra).
}
