% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sf.R
\name{get_geom_colname}
\alias{get_geom_colname}
\title{Get the geometry column name}
\usage{
get_geom_colname(x)
}
\arguments{
\item{x}{An sf object.}
}
\value{
A character.
}
\description{
Get the names of the geometry columns in an sf object.
}
