% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sf.R
\name{get_max}
\alias{get_max}
\title{Get maximum coordinates}
\usage{
get_max(x)
}
\arguments{
\item{x}{an sf or terra object.}
}
\value{
a named vector.
}
\description{
Get the maximum coordinates of the extent of the given  object (sf or
terra).
}
