% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sf.R
\name{get_min}
\alias{get_min}
\title{Get minimum coordinates}
\usage{
get_min(x)
}
\arguments{
\item{x}{an sf or terra object.}
}
\value{
a named vector.
}
\description{
Get the mininum coordinates of the extent of the given  object (sf or
terra).
}
