% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame.R
\name{listnames2dataframes}
\alias{listnames2dataframes}
\title{Add the list's names to the data frames in the list}
\usage{
listnames2dataframes(df_ls, cname)
}
\arguments{
\item{df_ls}{a named list of data frames.}

\item{cname}{a character(1). The name for the new column.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    a list of data frames.
}\if{html}{\out{</div>}}
}
\description{
Given a list of data frames, add each list's name to its corresponding data
frame.
}
