% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{make_grid}
\alias{make_grid}
\alias{make_grid_min_max_cells}
\alias{make_grid_origin_dist}
\alias{make_grid_origin_res}
\title{Build a grid that can be used later for aggregating geographic data}
\usage{
make_grid(
  x,
  n,
  cellsize = c(diff(sf::st_bbox(x)[c(1, 3)]), diff(sf::st_bbox(x)[c(2, 4)]))/n,
  id_col = "id",
  add_row_col = FALSE,
  add_centroids = FALSE,
  add_area = FALSE,
  drop_area_units = FALSE
)

make_grid_min_max_cells(
  xy_min,
  xy_max,
  n,
  crs = 4326,
  id_col = "id",
  add_row_col = FALSE,
  add_centroids = FALSE,
  add_area = FALSE,
  drop_area_units = FALSE
)

make_grid_origin_dist(
  xy_min,
  n,
  cell_size,
  crs,
  id_col = "id",
  add_row_col = FALSE,
  add_centroids = FALSE,
  add_area = FALSE,
  drop_area_units = FALSE
)

make_grid_origin_res(
  xy_origin,
  xy_min,
  xy_max,
  cell_size,
  crs = 4326,
  id_col = "id"
)
}
\arguments{
\item{x}{An sf or terra object.}

\item{n}{integer of lenght 1 or 2. Number of cells (columns, rows).}

\item{cellsize}{numeric of length 1 or 2 (width, length). It defaults to the
width and heigth of the bounding box of the given object divided by the
number of cells.}

\item{id_col}{character of length 1. Name of the column for identifying grid
cells.}

\item{add_row_col}{Either logical(1) or character(2). Should columns
identifying the row and column of each cell be added? If so, provide
their names; otherwise default names are used}

\item{add_centroids}{Either logical(1) or character(2). Should columns with
the coordinates of the cells' centroids be added? If so, provide their
names (X & Y); otherwise default names are used.}

\item{add_area}{Either logical(1) or character(1). Should a column with the
cell's area be added? If so, provide its name; otherwise a default name is
used.}

\item{drop_area_units}{A logical. Drop the area units?}

\item{xy_min, xy_max}{A pair of numeric vectors with the minimum and maximum
coordiantes of the grid.}

\item{crs}{Numeric or character representing a Coordinate Reference System.}

\item{cell_size}{A numeric of length 1 or 2. The width and heigth o a cell.
The units are those of the \code{crs}.}

\item{xy_origin}{A numeric(2) representing an XY point.}
}
\value{
An \code{sf} object with polygons.
}
\description{
Build a grid around spatial data (vector or raster) or from scratch.

\code{make_grid} returns a grid around spatial data (vector or raster).

\code{make_grid_min_max_cells} returns a grid using the minimum & maximum
coordinates, and the number of cells in the grid.

\code{make_grid_origin_dist} returns a grid using the coordinates of an origin
point (minimum XY), the number of cells, and the number of cells in the
grid.

\code{make_grid_origin_res} returns a grid using the coordinates of an origin
point inside an area and a cell size. The new grid will have a vertex on
the origin and will fit as many cells as possible in the area.
}
\examples{

# Make a 5x5 grid covering a polygon.
pol <- sf::st_sfc(sf::st_polygon(list(cbind(c(0, 3, 3, 0, 0), 
                                            c(0, 0, 3, 3, 0)))),
                  crs = 4326)
make_grid(pol, cellsize = 1, n = 5)


# Make a 5x10 grid covering a raster.
r <- terra::rast(matrix(1:25, nrow = 5, ncol = 5), crs = "EPSG:4326")
make_grid(r, cellsize = 1, n = c(10, 5))


# Make a grid using mininum and maximum coordinates and the number of grid 
# cells.
make_grid_min_max_cells(
    xy_min = c(-180, -90),
    xy_max = c(180, 90),
    n = c(14, 7),
    crs = 4326
) 


# Make a rectangular grid using square cells of 25 kilometers.
make_grid_origin_dist(
    xy_min = c(-74, 4),
    n = c(14, 7),
    cell_size = 25000,
    crs = 3116
) 


# Make a rectangular grid using square cells of 25 kilometers.
make_grid_origin_res(
    xy_origin = c(-74.2, 4.4),
    xy_min = c(-78, -3),
    xy_max = c(-70, 5),
    cell_size = 0.5
)

}
\seealso{
\code{\link[sf:st_make_grid]{sf::st_make_grid()}} which this function wraps.
}
