% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame.R
\name{merge_data_frames}
\alias{merge_data_frames}
\title{Merge a list of data frames}
\usage{
merge_data_frames(data_frames, by)
}
\arguments{
\item{data_frames}{a list of data.frame.}

\item{by}{a character of column names.}
}
\value{
a data.frame.
}
\description{
Merge a list of data.frame by a given column name.
}
