% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sf.R
\name{simplify_geoms}
\alias{simplify_geoms}
\title{Simplify geometries}
\usage{
simplify_geoms(data_sf, geom_types = NA)
}
\arguments{
\item{data_sf}{an sf object.}

\item{geom_types}{a character. Force specific geometry types. The first
element must be the name of a simple geometry (e.g. c("POINT",
"MULTIPOINT")).}
}
\value{
an sf object with simple geometries (i.e. POINT, LINESTRING,
POLYGON).
}
\description{
Convert geometries to simple parts. The geometries which can't be simplified
are discarded. This is useful when dealing with data resulting from spatial
operations such as intersection.
}
