% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeslabgrid-package.R
\docType{package}
\name{treeslabgrid-package}
\alias{treeslabgrid}
\alias{treeslabgrid-package}
\title{treeslabgrid: Create Grids and Use Them to Aggregate Data}
\description{
Convenience function to create regular grids and use them to aggregate geographic (vector or raster) data.
}
\author{
\strong{Maintainer}: Alber Sánchez \email{alber.ipia@inpe.br} (\href{https://orcid.org/0000-0001-7966-2880}{ORCID})

}
\keyword{internal}
